+++ src/tasks/index.ts
@@ -1,0 +1,1 @@
+export { TaskRegistry, type TaskEntry, type CreateTaskOpts } from "./registry.js";